requirements.txt
asgiref==3.3.0
Django==3.1.2
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.4.1
requests==2.24.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
